// web/src/api/schemas/config.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Litchi API
 * Automated development agent system - from GitHub Issue to Pull Request
 * OpenAPI spec version: 0.1.0
 */
import type { AgentConfig } from "./agentConfig";
import type { AuditConfigDTO } from "./auditConfigDTO";
import type { ClarityConfig } from "./clarityConfig";
import type { ComplexityConfig } from "./complexityConfig";
import type { DatabaseConfig } from "./databaseConfig";
import type { GitConfig } from "./gitConfig";
import type { ServerConfig } from "./serverConfig";
import type { WebhookConfig } from "./webhookConfig";

export interface Config {
	agent?: AgentConfig;
	audit?: AuditConfigDTO;
	clarity?: ClarityConfig;
	complexity?: ComplexityConfig;
	database?: DatabaseConfig;
	git?: GitConfig;
	server?: ServerConfig;
	webhook?: WebhookConfig;
}